# src/main/java/com/example/springindiv2/dto/AddressDto.java
package com.example.springindiv2.dto;

import lombok.Data;

@Data
public class AddressDto {
    private Integer id;
    private String street;
    private String city;
    private String bolge;
//    private String stateAndCity;
}